temporary solution to onnx-tensorrt bug
we cast text into float before argmax to avoid the bug in onnx-tensorrt

see https://github.com/onnx/onnx-tensorrt/issues/786

+++ clip/model.py
@@ -347,7 +347,11 @@ class CLIP(nn.Module):
 
         # x.shape = [batch_size, n_ctx, transformer.width]
         # take features from the eot embedding (eot_token is the highest number in each sequence)
-        x = x[torch.arange(x.shape[0]), text.argmax(dim=-1)] @ self.text_projection
+
+        # we cast text into float before argmax to avoid the bug in onnx-tensorrt
+        # see https://github.com/onnx/onnx-tensorrt/issues/786
+        # this casting should be removed once they fixed the bug
+        x = x[torch.arange(x.shape[0]), text.float().argmax(dim=-1)] @ self.text_projection
 
         return x
